Configure new plugin registry and broker
