+++ assembly/assembly-wsmaster-war/src/main/webapp/WEB-INF/classes/che/che.properties
@@ -551,8 +551,8 @@ che.singleport.wildcard_domain.ipless=false
 
 # Docker image of Che plugin broker app that resolves workspace tooling configuration and copies
 # plugins dependencies to a workspace
-che.workspace.plugin_broker.init.image=eclipse/che-init-plugin-broker:v0.16
-che.workspace.plugin_broker.unified.image=eclipse/che-unified-plugin-broker:v0.16
+che.workspace.plugin_broker.init.image=eclipse/che-init-plugin-broker:v0.17
+che.workspace.plugin_broker.unified.image=eclipse/che-unified-plugin-broker:v0.17
 
 # Docker image of Che plugin broker app that resolves workspace tooling configuration and copies
 # plugins dependencies to a workspace
@@ -564,7 +564,7 @@ che.workspace.plugin_broker.wait_timeout_min=3
 # Workspace tooling plugins registry endpoint. Should be a valid HTTP URL.
 # Example: http://che-plugin-registry-eclipse-che.192.168.65.2.nip.io
 # In case Che plugins tooling is not needed value 'NULL' should be used
-che.workspace.plugin_registry_url=https://che-plugin-registry.openshift.io/v2
+che.workspace.plugin_registry_url=https://che-plugin-registry.openshift.io/v3
 
 # Configures in which way secure servers will be protected with authentication.
 # Suitable values:
