+++ deploy/openshift/deploy_che.sh
@@ -204,7 +204,7 @@ export PLUGIN_REGISTRY_IMAGE=${PLUGIN_REGISTRY_IMAGE:-${DEFAULT_PLUGIN_REGISTRY_
 DEFAULT_PLUGIN_REGISTRY_IMAGE_PULL_POLICY="Always"
 export PLUGIN_REGISTRY_IMAGE_PULL_POLICY=${PLUGIN_REGISTRY_IMAGE_PULL_POLICY:-${DEFAULT_PLUGIN_REGISTRY_IMAGE_PULL_POLICY}}
 
-DEFAULT_PLUGIN__REGISTRY__URL="https://che-plugin-registry.openshift.io/v2"
+DEFAULT_PLUGIN__REGISTRY__URL="https://che-plugin-registry.openshift.io/v3"
 export PLUGIN__REGISTRY__URL=${PLUGIN__REGISTRY__URL:-${DEFAULT_PLUGIN__REGISTRY__URL}}
 
 DEFAULT_CHE_METRICS_ENABLED="false"
@@ -563,7 +563,7 @@ ${CHE_VAR_ARRAY}"
 
     if [ "${DEPLOY_CHE_PLUGIN_REGISTRY}" == "true" ]; then
         PLUGIN_REGISTRY_ROUTE=$($OC_BINARY get route/che-plugin-registry --namespace=${CHE_OPENSHIFT_PROJECT} -o=jsonpath={'.spec.host'})
-        PLUGIN__REGISTRY__URL="${HTTP_PROTOCOL}://${PLUGIN_REGISTRY_ROUTE}/v2"
+        PLUGIN__REGISTRY__URL="${HTTP_PROTOCOL}://${PLUGIN_REGISTRY_ROUTE}/v3"
     fi
 
     if [ ! -z ${CHE_INFRA_OPENSHIFT_PROJECT} ]; then
